+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## unreleased
 
+- Added support for local or remote files for any config file
+- Refactored the file handling
+- Added additional tests
+
 ## [1.15.3] - 2024-01-05
 
 - Bumped dependencies traverse and node-version of the Github actions